% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop-tif.R
\name{crop_worldpop_tif}
\alias{crop_worldpop_tif}
\title{crop_worldpop_tif}
\usage{
crop_worldpop_tif(city = "accra", expand = 0.1)
}
\arguments{
\item{city}{Name of city (one of "accra", "kathmandu", or "bristol")}

\item{expand}{Relative factor by which to expand cropping limits beyond
rectangular bounding box of city}
}
\value{
Nothing Files are written to the corresponding directories
}
\description{
\code{.zip} files downloaded from worldpop are national-scale. This function
crops them to the designated cities, and converts them to much smaller geotif
files.
}
\note{
This should be run from somewhere within the \code{whodata} repository
structure; anywhere else will likely fail.
}
