% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{get_osm_buildings}
\alias{get_osm_buildings}
\title{get_osm_buildings}
\usage{
get_osm_buildings(city = "kathmandu", n = 1)
}
\arguments{
\item{city}{Name of city for which buildings are to be obtained}

\item{n}{Number of chunks into which to divide the file (see details)}
}
\value{
The \pkg{sf}-formatted data object (invisibly)
}
\description{
Extract OSM buildings for given location (\code{city}), and save them in the
data directory
}
