% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdens_to_points.R
\name{pop2point}
\alias{pop2point}
\title{Assign raster cell values to points}
\usage{
pop2point(ras, nodes, redistribute_missing = NULL)
}
\arguments{
\item{ras}{Raster dataset from which cell values are taken}

\item{nodes}{Points onto which cells will be distributed}

\item{redistribute_missing}{How should empty cells be redistributed?
To the nearest node by default.}
}
\value{
A copy of \code{nodes} containing a new population variable (pop)
}
\description{
Assign raster cell values to points
}
\examples{
\dontrun{
devtools::install_github("robinlovelace/osmdata")
library(osmdata)
library(raster)
library(tmap)
library(sf)

boundary_bb = getbb("Accra")
boundary = stplanr::bb2poly(boundary_bb) \%>\% 
  sf::st_as_sf()
# from inside the who-data repo
old = setwd("~/ATFutures/who-data/")
ras = raster::raster("accra/popdens/GHA15_040213.tif")
setwd(old)
ras # 200k cells
# ras = mask(ras, as_Spatial(st_geometry(boundary)))
ras = crop(ras, extent(boundary_bb)) # make smaller dataset - 148k cells
ways = readRDS("../who-data/accra/osm/accra-hw.Rds")
nodes = sf::st_cast(ways, "POINT")
nodes_new = pop2point(ras, nodes)
}
}
