% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdens-package.R
\docType{package}
\name{popdens-package}
\alias{popdens-package}
\alias{_PACKAGE}
\title{popdens: Population density from Meta's "Data for Good" programme}
\description{
Population density from Meta's "Data for Good" programme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/popdens}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/popdens/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
