% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdens_to_points.R
\name{xy_to_sfc}
\alias{xy_to_sfc}
\title{xy_to_sfc}
\usage{
xy_to_sfc(xy)
}
\arguments{
\item{xy}{matrix, data.frame, or tibble of points}
}
\value{
sf::sfc representation of same
}
\description{
Convert matrix of xy points to sfc object
}
